release: v0.6.0 — Patch-06 conformance + docs + version bump
- crates/sccgub-node/tests/patch_06_conformance.rs: end-to-end scenario
  exercising all five v5 systems (§30 forgery-veto, §31 fee floor, §32
  fork-choice, §33 pruning, §34 upgrade) plus a replay-determinism test.
- Workspace version 0.5.0 → 0.6.0
- OpenAPI artifact regenerated to match new package version
- README.md: test count 1155 → 1209; v0.6.0 banner; Conformance Matrix
  extended with five new Patch-06 invariants
- docs/STATUS.md: test count 1155 → 1209; v5 capabilities line
- EXTERNAL_AUDIT_PREP.md: version 0.5.0 → 0.6.0; test count 1155 → 1209
- CHANGELOG.md: full Patch-06 entry under [v0.6.0]

Workspace posture: 1209 tests via `cargo test -- --list`, cargo fmt
clean, cargo clippy --workspace --all-targets -D warnings clean.

+++ crates/sccgub-execution/src/chain_version_check.rs
@@ -65,7 +65,7 @@ pub fn verify_block_version_alignment(
             break;
         }
     }
-    if active_version as u32 == block_version {
+    if active_version == block_version {
         ChainVersionCheck::Aligned
     } else {
         ChainVersionCheck::Misaligned(ChainVersionRejection::VersionOutOfAlignment {
